+++ routes/account.js
@@ -1,16 +1,18 @@
 var path = require('path')
     ,konphyg = require('konphyg')(path.join(__dirname, '../config'))
     ,express_cfg = konphyg('express')
-    ,api = require(express_cfg.api);
-
+    ,api = require(express_cfg.api)
+    ,lib = require('../lib');
 
 exports.index = function(req, res){
+    var data = api.account.get();
+    lib.templating.amend_json(data);
     res.format({
         json: function(){
-            res.json(api.account.get());
+            res.json(data);
         },
         html: function(){
-            res.render('account', api.account.get());
+            res.render('account', data);
         }
     })
 };