+++ routes/errors.js
@@ -1,7 +1,7 @@
 var lib = require('../lib');
 
 exports.page_not_found = function(req, res){
-    console.log('- error not found');
+    console.log(' - error not found');
     res.statusCode = 404;
     res.end();
     return;