+++ routes/github.js
@@ -1,17 +1,18 @@
 var path = require('path')
     ,konphyg = require('konphyg')(path.join(__dirname, '../config'))
     ,express_cfg = konphyg('express')
-    ,api = require(express_cfg.api);
-
-
+    ,api = require(express_cfg.api)
+    ,lib = require('../lib');
 
 exports.index = function(req, res){
+    var data = api.github.get();
+    lib.templating.amend_json(data);
     res.format({
         json: function(){
-            res.json(api.github.get());
+            res.json(data);
         },
         html: function(){
-            res.render('github', api.github.get());
+            res.render('github', data);
         }
     })
 };